Updates for Heroku production mode
Makes changes to the basic project structure to support local
development with hot module reloading from webpack, as well as
production server deployment, intended for Heroku.

Effectively, we create a proxy to spin up in development mode, and route
traffic through.  This seamlessly maps to a development webpack server
when we're in production mode.  When not, it's totally skipped, and we
are instead just running an express server.

We add a postinstall script, which in production mode, does a production
build process.  This ensures that we can run the same basic commands,
and get the environment appropriate results.

There's no need for a Procfile, because Heroku will assume you want to
run 'npm start' if you don't provide one for a Node.js application,
which this basic app is.

We update gitignore to exclude npm debug logs, and move a few files
around.

The bulk of this change was influenced by [The Ultimate Webpack
Setup](http://www.christianalfoni.com/articles/2015_04_19_The-ultimate-webpack-setup).
I had to tweak a few things, but it mostly follows along.